fix extra space when description is empty
